backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic[email]